End- Algo & Error Handling

+++ app/controller/patientsController.js
@@ -12,12 +12,18 @@ const patientsController = {
      */
     async getAll(_,res,next) {
         try {
-        const allPatients= await patientsDatamapper.findAll();
-        res.json(allPatients)
-        } catch {
-            next(new APIError("Erreur lors de la récupération des patients", 500));
+            const allPatients= await patientsDatamapper.findAll();
+                
+            if (allPatients.length === 0) {
+                next(new APIError("La route n'a pas été trouvé", 404));
+            } else {
+                res.json(allPatients);  
+            }
+            } catch {
+                next(new APIError("Erreur lors de la récupération des patients", 500));
         }
     },
+
     /**
      * Récupération d'un patient par son id
      * @param {*}req requête Express
@@ -27,12 +33,19 @@ const patientsController = {
     async getById(req,res,next) {
         const id = req.params.id
         try {
-        const onePatientsById = await patientsDatamapper.findByPk(id);
-        res.json(onePatientsById)
-        } catch {
-            next(new APIError("Erreur lors de la récupération du patient", 500));
+            const onePatientsById = await patientsDatamapper.findByPk(id);
+            
+            if (onePatientsById.length === 0) {
+                next(new APIError("La route n'a pas été trouvé", 404));
+
+            } else {
+                res.json(onePatientsById)
+            }
+            } catch {
+                    next(new APIError("Erreur lors de la récupération du patient", 500));
         }
     },
+    
     // dans le patientsInfo il y aura quizz_id
     //et le quizz id sera recupéré soit via le body ou params
     /**
@@ -42,7 +55,7 @@ const patientsController = {
      * @returns {json} créer un patient 
      */
     async createPatients(req,res,next) {
-        try {
+        
         const patientsInfo = {
             email : req.body.email,
             lastname: req.body.lastname,
@@ -54,10 +67,20 @@ const patientsController = {
             city : req.body.city,
             quizz_id : req.body.quizz_id     
         }
-        const createPatients = await patientsDatamapper.create(patientsInfo);
-        res.json(createPatients)
-        } catch {
-        next(new APIError("Erreur lors de la création du patient", 500));
+
+        if (!patientsInfo){
+            next(new APIError("Paramètres manquants", 400));
+        }
+        try {
+            const createPatients = await patientsDatamapper.create(patientsInfo);
+            
+            if (createPatients.length === 0) {
+                next(new APIError("La route n'a pas été trouvé", 404));
+            } else {
+                res.json(createPatients)
+            }
+            } catch {
+            next(new APIError("Erreur lors de la création du patient", 500));
         }
     },
     /**
@@ -68,7 +91,11 @@ const patientsController = {
      */
     async updatePatients(req,res,next){
         const id = req.params.id
-        try {
+        
+        if (!id){
+            next(new APIError("Paramètres manquants", 400));
+        }
+
         const patientsInfo = {
             email : req.body.email,
             lastname: req.body.lastname,
@@ -80,11 +107,22 @@ const patientsController = {
             city : req.body.city,
             role : req.body.role
         }
-        const updatePatients = await patientsDatamapper.update({id},patientsInfo);
-        res.json(updatePatients)
-        } catch {
-            next(new APIError("Erreur lors de la mise à jour du patient", 500));
+        
+        if (!patientsInfo){
+            next(new APIError("Paramètres manquants", 400));
         }
+        
+        try {
+            const updatePatients = await patientsDatamapper.update({id},patientsInfo);
+            
+            if (updatePatients.length === 0) {
+                next(new APIError("La route n'a pas été trouvé", 404));
+            } else {
+                res.json(updatePatients)
+            }
+            } catch {
+                next(new APIError("Erreur lors de la mise à jour du patient", 500));
+            }
     },
     /**
      * Suppression d'un patient
@@ -94,12 +132,23 @@ const patientsController = {
      */
     async deletePatients(req,res,next){
         const id = req.params.id
-        try {
-        const deletePatients = await patientsDatamapper.delete(id);
-        res.json(deletePatients)
-        } catch {
-            next(new APIError("Erreur lors de la suppression du patient", 500));
+
+        if (!id){
+            next(new APIError("Paramètres manquants", 400));
+            return;
         }
+        try {
+            const deletePatients = await patientsDatamapper.delete(id);
+            
+        if (deletePatients.length === 0) {
+            next(new APIError("La route n'a pas été trouvé", 404));
+
+        } else {
+            res.json(deletePatients)
+        }
+        }catch {
+                next(new APIError("Erreur lors de la suppression du patient", 500));
+            }
     },
     /**
      * Récupération d'un patient avec ses rendez-vous
@@ -109,12 +158,24 @@ const patientsController = {
      */
     async getOnePatientWithAllAppointments (req,res,next) {
         const id = req.params.id
-        try {
-        const getOnePatientsWithAllAppointments = await patientsDatamapper.getOnePatientWithAllAppointments(id);
-        res.json(getOnePatientsWithAllAppointments)
-        } catch {
-            next(new APIError("Erreur lors de la récupération du patient avec ses rendez-vous", 500));
+
+        if (!id){
+            next(new APIError("Paramètres manquants", 400));
+            return;
         }
+
+        try {
+            const getOnePatientsWithAllAppointments = await patientsDatamapper.getOnePatientWithAllAppointments(id);
+            
+            if (getOnePatientsWithAllAppointments.length === 0) {
+                next(new APIError("La route n'a pas été trouvé", 404));
+
+            } else {
+                res.json(getOnePatientsWithAllAppointments)
+            }
+            } catch {
+                next(new APIError("Erreur lors de la récupération du patient avec ses rendez-vous", 500));
+            }
     },
     /**
      * Récupération d'un patient avec ses quizz
@@ -124,12 +185,24 @@ const patientsController = {
      */
     async getOnePatientWithQuizz (req,res,next) {
         const id= req.params.id
-        try {
-        const getOnePatientWithQuizz = await patientsDatamapper.getOnePatientWithQuizz(id);
-        res.json(getOnePatientWithQuizz)
-        } catch {
-            next(new APIError("Erreur lors de la récupération du patient avec ses quizz", 500));
+
+        if (!id){
+            next(new APIError("Paramètres manquants", 400));
+            return;
         }
+
+        try {
+            const getOnePatientWithQuizz = await patientsDatamapper.getOnePatientWithQuizz(id);
+            
+            if (getOnePatientWithQuizz.length === 0) {
+                next(new APIError("La route n'a pas été trouvé", 404));
+
+            } else {
+                res.json(getOnePatientWithQuizz)
+            }
+            } catch {
+                next(new APIError("Erreur lors de la récupération du patient avec ses quizz", 500));
+            }
     },
     /**
      * Récupération l'avis des patients sur un therapist
@@ -139,12 +212,23 @@ const patientsController = {
      */
     async getReviewsOneTherapists (req,res,next){
         const id= req.params.id
-        try {
-        const getReviewsOneTherapists = await patientsDatamapper.getReviewsOneTherapists(id);
-        res.json(getReviewsOneTherapists)
-        } catch {
-            next(new APIError("Erreur lors de la récupération des avis des patients sur un therapist", 500));
+
+        if (!id){
+            next(new APIError("Paramètres manquants", 400));
+            return;
         }
+        try {
+            const getReviewsOneTherapists = await patientsDatamapper.getReviewsOneTherapists(id);
+            
+            if (getReviewsOneTherapists.length === 0) {
+                next(new APIError("La route n'a pas été trouvé", 404));
+
+            }else {
+                res.json(getReviewsOneTherapists)
+            }
+            } catch {
+                next(new APIError("Erreur lors de la récupération des avis des patients sur un therapist", 500));
+            }
     },
     /**
      * création d'un rendez-vous entre un patient et un therapist
@@ -156,8 +240,16 @@ const patientsController = {
         
         const therapistId = req.params.therapistId
         const patientId= req.params.patientId
-        try {
         
+        if (!therapistId){
+            next(new APIError("Paramètres manquants", 400));
+            return;
+        }
+         if (!patientId){
+            next(new APIError("Paramètres manquants", 400));
+            return;
+        }
+
         const appointment = {
             beginninghour: req.body.beginninghour,
             endtime: req.body.endtime,
@@ -168,11 +260,22 @@ const patientsController = {
             chatsession : req.body.chatsession,
             sessionatoffice : req.body.sessionatoffice,
         }
-        
-        const createAppointmentOneTherapist = await patientsDatamapper.createAppointmentOneTherapist({therapistId,patientId},appointment);
-        res.json(createAppointmentOneTherapist)
-        } catch {
-            next(new APIError("Erreur lors de la création du rendez-vous entre un patient et un therapist", 500));
+
+        if (!appointment){
+            next(new APIError("Paramètres manquants", 400));
+            return;
+        }
+        try {
+            const createAppointmentOneTherapist = await patientsDatamapper.createAppointmentOneTherapist({therapistId,patientId},appointment);
+            
+            if (createAppointmentOneTherapist.length === 0) {
+                next(new APIError("La route n'a pas été trouvé", 404));
+
+            } else {
+                res.json(createAppointmentOneTherapist)
+            }
+            } catch {
+                next(new APIError("Erreur lors de la création du rendez-vous entre un patient et un therapist", 500));
         }
     },
     /**
@@ -184,7 +287,16 @@ const patientsController = {
     async createReviewsOneTherapist (req,res,next){
         const patientId= req.params.patientId
         const therapistId = req.params.therapistId
-        try {
+        
+        if (!patientId){
+            next(new APIError("Paramètres manquants", 400));
+            return;
+        }
+
+        if (!therapistId){
+            next(new APIError("Paramètres manquants", 400));
+            return;
+        }
         
         const review = {
             messages: req.body.messages,
@@ -193,11 +305,23 @@ const patientsController = {
             patients_id:patientId,
             therapists_id:therapistId,
         }
-        const createReviewsOneTherapist = await patientsDatamapper.createReviewsOneTherapist({patientId,therapistId},review);
-        res.json(createReviewsOneTherapist)
-        } catch {
-            next(new APIError("Erreur lors de la création d'un avis sur un therapist", 500));
+
+        if (!review){
+            next(new APIError("Paramètres manquants", 400));
+            return;
         }
+        try {
+            const createReviewsOneTherapist = await patientsDatamapper.createReviewsOneTherapist({patientId,therapistId},review);
+            
+            if (createReviewsOneTherapist.length === 0) {
+                next(new APIError("La route n'a pas été trouvé", 404));
+
+            }else {
+                res.json(createReviewsOneTherapist)
+            }
+            } catch {
+                next(new APIError("Erreur lors de la création d'un avis sur un therapist", 500));
+            }
     },  
     /**
      * Répondre au quizz pour récupérer le quizz_id
@@ -207,9 +331,7 @@ const patientsController = {
      */
     async answerPatientsQuizz(req,res,next) {
         
-        try {
-
-        const answers = {
+            const answers = {
             answer_1 : req.body.answer_1,
             answer_2 : req.body.answer_2,
             answer_3 : req.body.answer_3,
@@ -230,12 +352,24 @@ const patientsController = {
             answer_18 : req.body.answer_18,
 
         }
-        const answerPatientsQuizz = await patientsDatamapper.answerPatientsQuizz(answers);
-        res.json(answerPatientsQuizz)
-        } catch {
-             next(new APIError("Erreur lors de la réponse au quizz", 500));
-        
+
+        if (!answers){
+            next(new APIError("Paramètres manquants", 400));
+            return;
         }
+
+        try {
+            const answerPatientsQuizz = await patientsDatamapper.answerPatientsQuizz(answers);
+            
+            if (answerPatientsQuizz.length === 0) {
+                next(new APIError("La route n'a pas été trouvé", 404));
+            }else {
+                res.json(answerPatientsQuizz)
+            }
+            } catch {
+                next(new APIError("Erreur lors de la réponse au quizz", 500));
+            
+            }
     }
